pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cra-dexterous"
version = "0.1.0"
description = "Cascaded Residual Adaptation for Dexterous Manipulation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tensorboard>=2.14.0",
]

[project.optional-dependencies]
analysis = [
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
]

[tool.setuptools.packages.find]
include = ["cra*"]
